- Updated build and publish scripts to support specifying which project branch container profile to build and publish

+++ tools/util/create-and-publish-project-fs.sh
@@ -3,7 +3,7 @@
 dir=`dirname $0`
 
 # Set up the path option
-options=$(getopt -l profile:,version: -- "$@")
+options=$(getopt -o cp -l profile:,version:,create,publish -- "$@")
 
 if [ $? -ne 0 ]; then
     exit 1
@@ -11,9 +11,14 @@ fi
 
 eval set -- "$options"
 
+create=1
+publish=1
+
 # Fetch command line parameters
 while true; do
-    case "$1" in
+	case "$1" in
+        -c|--create)    create=1; publish=0; shift 1;;
+        -p|--publish)   publish=1; create=0; shift 1;;
         --profile)      profile=$2; shift 2;;
         --version)      version=$2; shift 2;;
         *)              break ;;
@@ -30,16 +35,20 @@ if [ -z "$version" ]; then
     exit 1
 fi
 
-"$dir/create-golden-project-container.sh"
+if [ "$create" -eq 1 ]; then
+	"$dir/create-golden-project-container.sh"
+fi
 
-tar --numeric-owner -C /var/lib/lxc/_golden-proj_/rootfs -czf /var/lib/lxc/_golden-proj_/rootfs.tgz .
+if [ "$publish" -eq 1 ]; then
+	tar --numeric-owner -C /var/lib/lxc/_golden-proj_/rootfs -czf /var/lib/lxc/_golden-proj_/rootfs.tgz .
 
-major=`echo $version | cut -d'.' -f1`
-upload_path="/var/www/nginx/rootfs/project/$profile/$major/project-$profile-$version.tgz"
-upload_dir=`dir $upload_path`
+	major=`echo $version | cut -d'.' -f1`
+	upload_path="/var/www/nginx/rootfs/project/$profile/$major/project-$profile-$version.tgz"
+	upload_dir=`dir $upload_path`
 
-ssh rainmaker@image.rainmaker.localdev "mkdir -p $upload_dir"
-scp /var/lib/lxc/_golden-proj_/rootfs.tgz rainmaker@image.rainmaker.localdev:"$upload_path"
+	ssh rainmaker@image.rainmaker.localdev "mkdir -p $upload_dir"
+	scp /var/lib/lxc/_golden-proj_/rootfs.tgz rainmaker@image.rainmaker.localdev:"$upload_path"
 
-unlink /var/lib/lxc/_golden-proj_/rootfs.tgz
-lxc-destroy -n _golden-proj_
+	unlink /var/lib/lxc/_golden-proj_/rootfs.tgz
+	lxc-destroy -n _golden-proj_
+fi
